requirements.txt
pyodbc>=4.0.39
python-dotenv>=1.0.0
markdown>=3.5.0
langgraph>=0.0.30
langchain>=0.3.27
langchain-groq>=0.1.0
langchain-core>=0.1.0
schedule>=1.2.0
pytz>=2024.1
